Remove box from composer add Symfony console
Remove box so that composer update --no-dev in the
Makefile gets rid of all unnecessary dependencies
before building the phar file. Phar size is now
only 2.3mb compared to 23mb. Makefile expects
box.phar to be locally installed, but this is not a
problem as not many contributors will be building
phars for now. Switch to phive and phpab probably
a good idea.