Upgrade PHP to 7.2 and PHPUnit to 8.0 :arrow_up:
Remove null division check because DependencyMap does not allow nodes
without any in or outgoing dependencies. So total dependencies for will
never be 0.

Fix some Docker issues when running PHPUnit tests in the Docker
container, because the user has root permissions and tests checking for
files which are not writable or readable fail. Solved using vfsstream
which is the better solution for unit tests anyways.